JUNE/requirements.txt
anthropic>=0.34.0
openai>=1.40.0
google-generativeai>=0.7.0
httpx>=0.27.0
playwright>=1.45.0
google-auth>=2.30.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.140.0
chromadb>=0.5.0
cryptography>=42.0.0
python-dotenv>=1.0.0
pywhatkit>=5.4
twilio>=9.2.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
openai-whisper>=20231117
rich>=13.7.0
fastapi>=0.111.0
uvicorn>=0.30.0
